Iteration 2: recipe page
Create a new directory within the odin-recipes directory and name it recipes. Create a new HTML file within the recipes directory and name it after the recipe it will contain. For now, just include an h1 heading with the recipe’s name as its content. Back in the index.html file, add a link to the recipe page you just created.